fix "warning: subscript has type `char'"
This patch fixes the following warning from GCC 3.4.2 on Solaris by
adding an appropriate cast (to int) for isspace:

bitconvert.c:197: warning: subscript has type `char'

+++ bitconvert.c
@@ -194,7 +194,7 @@ int bc_decode_track_fields(char* input, int encoding, int track, FILE* formats,
 	temp_ptr++;
 
 	/* skip whitespace between ':' and regular expression */
-	while (isspace(temp_ptr[0]) && temp_ptr[0] != '\0') {
+	while (isspace((int)temp_ptr[0]) && temp_ptr[0] != '\0') {
 		temp_ptr++;
 	}
